Massive Sunday overhaul
This should have been in more commits, but it's Sunday and I'm lazy

Way more agnostic about bots and less of a focus on security for local
run bots. Remote bots are the ultimate solution, and now we have one
running on Heroku for testing.

Next up, using Q on bot creation to ease setup